cdn/requirements.txt
flask == 3.1.0
pillow == 11.0.0
python-dotenv == 1.0.1